requirements.txt
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
logging==0.4.9.6
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
